% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.bootstrap_inference}
\alias{.bootstrap_inference}
\title{Inference for Nonlinear DiD}
\usage{
.bootstrap_inference(
  attgt_df,
  data,
  yname,
  tname,
  idname,
  gname,
  xformla,
  outcome_model,
  estimand,
  control_group,
  doubly_robust,
  nboot,
  boot_type,
  alpha,
  anticipation,
  parallel,
  pl_cores,
  data_type = "panel",
  cluster_var = NULL
)
}
\description{
Internal functions for bootstrap and delta-method standard
  errors in nonlinear staggered DiD, for both panel and repeated
  cross-section designs, with optional clustering and sampling weights.
}
\keyword{internal}
