% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiplier_weights_cpp}
\alias{multiplier_weights_cpp}
\title{Fast multiplier bootstrap weights}
\usage{
multiplier_weights_cpp(n, nboot)
}
\arguments{
\item{n}{Integer. Number of units.}

\item{nboot}{Integer. Number of bootstrap iterations.}
}
\value{
A matrix of dimension n x nboot with positive weights
that sum to n in each column.
}
\description{
Generates exponential multiplier weights for the
wild/multiplier bootstrap used in nonlinear DiD inference.
}
\keyword{internal}
