% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nonlineardid_methods}
\alias{nonlineardid_methods}
\title{S3 Methods for NonlinearDiD Objects}
\description{
Print, summary, and plot methods for \code{nonlinear_attgt}
  and \code{nonlinear_aggte} objects.
}
