% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.nonlinear_aggte}
\alias{plot.nonlinear_aggte}
\title{Plot Aggregated DiD Estimates}
\usage{
\method{plot}{nonlinear_aggte}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nonlinear_aggte}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots event-study, group-level, calendar, or overall
aggregated ATT estimates from \code{\link{nonlinear_aggte}}.
}
