% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean_cpp}
\alias{weighted_mean_cpp}
\title{Weighted mean (fast)}
\usage{
weighted_mean_cpp(x, w)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Numeric weights vector.}
}
\value{
Weighted mean scalar.
}
\description{
Computes weighted mean efficiently for bootstrap.
}
\keyword{internal}
